test: fix delete user errors

+++ server/src/use-cases/user/delete-user.js
@@ -6,6 +6,10 @@ export class DeleteUserUseCase {
     async execute(userId) {
         const deletedUser = await this.postgresDeleteUserRepository.execute(userId);
 
+        if (!deletedUser) {
+            throw new Error('User do not exists');
+        }
+
         await this.clerkClientAdapter.deleteUser(deletedUser.external_id);
 
         return deletedUser;
